+++ sticker/__init__.py
@@ -4,6 +4,7 @@ Matrix Sticker 模块
 提供 sticker 消息组件、存储管理和发送功能
 """
 
+from .availability import StickerAvailabilityStore
 from .component import Sticker, StickerInfo
 from .storage import StickerStorage
 from .syncer import StickerPackInfo, StickerPackSyncer
@@ -12,6 +13,7 @@ __all__ = [
     "Sticker",
     "StickerInfo",
     "StickerStorage",
+    "StickerAvailabilityStore",
     "StickerPackSyncer",
     "StickerPackInfo",
 ]